Turn `UploadArea` into `FileInput`
The `value` is no longer an AsyncData but a union that's either a
`File` or a `DistFile` (consisting of a `url` and an optional `name`).

While the previous contract forced the consumer to provide a File in
order to get a proper mimeType (to know if an image preview was
displayable), we now use a heuristic based on the provided `accept`
type: if the only uploadable mimeTypes are images, we show the preview,
otherwise we default to an icon.

Also renamed onDropAccepted -> onFiles and onDropRejected -> onError for
clarity