Add `useFilesUploader` hook.
This hook provides a generic interface for our document upload use-cases
(SupportingDocumentCollection & CapitalDepositCase), where you can
upload multiple files.

For those, the process is the following:
-> Call a mutation to generate the upload "configuration" (url, fields
to send in the body)
-> Effectively upload the file to the storage service

At the moment, the storage service doesn't notify us when the file is
effectively processed, meaning we need to rely on the local information.

This hooks takes as config:
-> `initialFiles`, which we receive from the API
-> `generateUpload`, to get the upload configuration (its type is
generic to account for all cases). Note that the document has an ID only
once this operation is successful.
-> `uploadFile`, to send the file to the storage service
-> `onError`, called if one of the two steps fails

It then manages the file lifecycle within its state.

Known limitation: if new `files` are received from the API, we don't
reconcile. This shouldn't be an issue.