Revamp Supporting Documents Collection (#325)
* Add `SwanFile` type to share between interface elements

* Turn `UploadArea` into `FileInput`

The `value` is no longer an AsyncData but a union that's either a
`File` or a `DistFile` (consisting of a `url` and an optional `name`).

While the previous contract forced the consumer to provide a File in
order to get a proper mimeType (to know if an image preview was
displayable), we now use a heuristic based on the provided `accept`
type: if the only uploadable mimeTypes are images, we show the preview,
otherwise we default to an icon.

Also renamed onDropAccepted -> onFiles and onDropRejected -> onError for
clarity

* Add `useFilesUploader` hook.

This hook provides a generic interface for our document upload use-cases
(SupportingDocumentCollection & CapitalDepositCase), where you can
upload multiple files.

For those, the process is the following:
-> Call a mutation to generate the upload "configuration" (url, fields
to send in the body)
-> Effectively upload the file to the storage service

At the moment, the storage service doesn't notify us when the file is
effectively processed, meaning we need to rely on the local information.

This hooks takes as config:
-> `initialFiles`, which we receive from the API
-> `generateUpload`, to get the upload configuration (its type is
generic to account for all cases). Note that the document has an ID only
once this operation is successful.
-> `uploadFile`, to send the file to the storage service
-> `onError`, called if one of the two steps fails

It then manages the file lifecycle within its state.

Known limitation: if new `files` are received from the API, we don't
reconcile. This shouldn't be an issue.

* Make `FileTile` more versatile

-> Uses `SwanFile` values
-> Handles `Uploading` status
-> Embeds its own translations for file statuses
-> Handles deletion loading state

* Add `FilesUploader` component

Wraps the newly introduced `useFilesUploader` hook, providing:
-> `getUploadConfig` to allow for genericity
-> `onRemoveFile` to allow file deletion
-> `onChange` to keep track of the updates

* Add @swan-io/request

* Add SupportingDocumentCollection

Handles Supporting Document Collection cases for all situations in the
dashboard or the end-customer exposed interfaces

* Remove old SupportingDocument component

* Format file

* Cleanup translations

* Rename `onboardingLanguage` to `templateLanguage`

* Add translations